# perf: Optimize cloning of Context since it is immutable