--- a/opentelemetry/src/context.rs
+++ b/opentelemetry/src/context.rs
@@ -93,11 +93,12 @@ thread_local! {
 /// assert_eq!(current.get::<ValueB>(), None);
 /// ```
 #[derive(Clone, Default)]
+#[repr(C)]
 pub struct Context {
     #[cfg(feature = "trace")]
     pub(crate) span: Option<Arc<SynchronizedSpan>>,
     entries: Option<Arc<EntryMap>>,
-    suppress_telemetry: bool,
+    flags: ContextFlags,
 }
 
 type EntryMap = HashMap<TypeId, Arc<dyn Any + Sync + Send>, BuildHasherDefault<IdHasher>>;
@@ -245,7 +246,7 @@ impl Context {
             entries,
             #[cfg(feature = "trace")]
             span: self.span.clone(),
-            suppress_telemetry: self.suppress_telemetry,
+            flags: self.flags,
         }
     }
 
@@ -335,7 +336,7 @@ impl Context {
     /// Returns whether telemetry is suppressed in this context.
     #[inline]
     pub fn is_telemetry_suppressed(&self) -> bool {
-        self.suppress_telemetry
+        self.flags.is_telemetry_suppressed()
     }
 
     /// Returns a new context with telemetry suppression enabled.
@@ -344,7 +345,7 @@ impl Context {
             entries: self.entries.clone(),
             #[cfg(feature = "trace")]
             span: self.span.clone(),
-            suppress_telemetry: true,
+            flags: self.flags.with_telemetry_suppressed(),
         }
     }
 
@@ -413,7 +414,7 @@ impl Context {
         Self::map_current(|cx| Context {
             span: Some(Arc::new(value)),
             entries: cx.entries.clone(),
-            suppress_telemetry: cx.suppress_telemetry,
+            flags: cx.flags,
         })
     }
 
@@ -422,7 +423,7 @@ impl Context {
         Context {
             span: Some(Arc::new(value)),
             entries: self.entries.clone(),
-            suppress_telemetry: self.suppress_telemetry,
+            flags: self.flags,
         }
     }
 }
@@ -446,7 +447,7 @@ impl fmt::Debug for Context {
         let entries = self.entries.as_ref().map_or(0, |e| e.len());
 
         dbg.field("entries count", &entries)
-            .field("suppress_telemetry", &self.suppress_telemetry)
+            .field("suppress_telemetry", &self.flags.is_telemetry_suppressed())
             .finish()
     }
 }
@@ -502,6 +503,7 @@ impl Hasher for IdHasher {
 /// [`ContextGuard`] instances that are constructed using ids from it can't be
 /// moved to other threads. That means that the ids are always valid and that
 /// they are always within the bounds of the stack.
+#[repr(C)]
 struct ContextStack {
     /// This is the current [`Context`] that is active on this thread, and the top
     /// of the [`ContextStack`]. It is always present, and if the `stack` is empty
@@ -602,6 +604,23 @@ impl Default for ContextStack {
             stack: Vec::with_capacity(ContextStack::INITIAL_CAPACITY),
             _marker: PhantomData,
         }
+    }
+}
+
+#[derive(Clone, Copy, Default)]
+struct ContextFlags(usize);
+
+impl ContextFlags {
+    const TELEMETRY_SUPPRESSED: usize = 1;
+
+    #[inline(always)]
+    fn is_telemetry_suppressed(&self) -> bool {
+        self.0 & Self::TELEMETRY_SUPPRESSED != 0
+    }
+
+    #[inline(always)]
+    fn with_telemetry_suppressed(&self) -> Self {
+        Self(self.0 | Self::TELEMETRY_SUPPRESSED)
     }
 }