This PR is just for discussion. It's built on top of #2821 and the changes are in this commit 
[dc038f9](https://github.com/open-telemetry/opentelemetry-rust/pull/2861/commits/dc038f9389b23e7925c6de4131565f30773195b2)

These are the results for the benchmarks on my M1 Max 

| Benchmark                             | Base    | Changed |
|---------------------------------------|---------|---------|
| enter_telemetry_suppressed_scope      | 17.3 ns | 26.9 ns |
| normal_attach                         | 17.3 ns |  8.1 ns |
| is_current_telemetry_suppressed_false |  1.1 ns |  1.6 ns |
| is_current_telemetry_suppressed_true  |  1.1 ns |  1.8 ns |

So creating a new context in `enter_telemetry_suppressed_scope` is more expensive because of the extra `Arc`, but on the other hand `normal_attach` is way faster because the `clone` is now cheap.

The performance in `context_attach` is now 12%-22% faster than `main` and 44%-63% faster than the context suppression branch.