@scottgerring 2025-03-26
👷 build bot output from [this run](https://github.com/open-telemetry/opentelemetry-rust/actions/runs/14082168546/job/39437405545?pr=2861): 

| name                                                  | baseDuration   | changesDuration   | difference |
|-------------------------------------------------------|---------------|------------------|------------|
| context_attach/nested_cx/empty_cx                    | 31.7±0.14ns   | **19.1±0.27ns**  | -40        |
| context_attach/nested_cx/single_value_cx             | 33.7±0.22ns   | **19.4±0.44ns**  | -42        |
| context_attach/nested_cx/span_cx                     | 33.0±0.29ns   | **19.4±0.30ns**  | -41        |
| context_attach/out_of_order_cx_drop/empty_cx         | 38.7±0.18ns   | **19.8±0.19ns**  | -49        |
| context_attach/out_of_order_cx_drop/single_value_cx  | 40.1±0.13ns   | **20.8±0.48ns**  | -48        |
| context_attach/out_of_order_cx_drop/span_cx          | 39.7±0.22ns   | **20.8±0.32ns**  | -48        |
| context_attach/single_cx/empty_cx                    | 18.1±0.22ns   | **12.5±0.31ns**  | -31        |
| context_attach/single_cx/single_value_cx             | 17.6±0.08ns   | **12.5±0.13ns**  | -29        |
| context_attach/single_cx/span_cx                     | 17.3±0.13ns   | **12.5±0.25ns**  | -28        |