@bantonsson 2025-03-27
These are the benchmark numbers from [this run](https://github.com/open-telemetry/opentelemetry-rust/actions/runs/14101769547/job/39499330671?pr=2861):

And yes, the suppression check is slower, but on the other hand context attach and cloning is way faster.

| name                                                          | baseDuration         | changesDuration      | difference |
|---------------------------------------------------------------|----------------------|----------------------|------------|
| `context/has_active_span/in-cx/alt`                           | `8.4±0.01ns`         | `8.4±0.02ns`         | `0.0`      |
| `context/has_active_span/in-cx/spec`                          | `4.4±0.06ns`         | `4.4±0.07ns`         | `0.0`      |
| `context/has_active_span/no-cx/alt`                           | `8.4±0.02ns`         | `8.4±0.01ns`         | `0.0`      |
| `context/has_active_span/no-cx/spec`                          | `4.4±0.09ns`         | `4.4±0.05ns`         | `0.0`      |
| `context/has_active_span/no-sdk/alt`                          | `8.4±0.03ns`         | `8.4±0.01ns`         | `0.0`      |
| `context/has_active_span/no-sdk/spec`                         | `4.4±0.10ns`         | `4.4±0.06ns`         | `0.0`      |
| `context/is_recording/in-cx/alt`                              | `4.7±0.06ns`         | `4.7±0.06ns`         | `0.0`      |
| `context/is_recording/in-cx/spec`                             | `6.6±0.08ns`         | `6.6±0.07ns`         | `0.0`      |
| `context/is_recording/no-cx/alt`                              | `4.7±0.05ns`         | `4.7±0.04ns`         | `0.0`      |
| `context/is_recording/no-cx/spec`                             | `6.5±0.06ns`         | `6.6±0.15ns`         | `0.0`      |
| `context/is_recording/no-sdk/alt`                             | `4.7±0.07ns`         | `4.7±0.07ns`         | `0.0`      |
| `context/is_recording/no-sdk/spec`                            | `6.5±0.07ns`         | `6.5±0.09ns`         | `0.0`      |
| `context/is_sampled/in-cx/alt`                                | `8.7±0.02ns`         | `8.7±0.03ns`         | `0.0`      |
| `context/is_sampled/in-cx/spec`                               | `4.6±0.06ns`         | `4.6±0.07ns`         | `0.0`      |
| `context/is_sampled/no-cx/alt`                                | `8.7±0.04ns`         | `8.7±0.02ns`         | `0.0`      |
| `context/is_sampled/no-cx/spec`                               | `4.7±0.06ns`         | `4.7±0.09ns`         | `0.0`      |
| `context/is_sampled/no-sdk/alt`                               | `8.7±0.03ns`         | `8.7±0.03ns`         | `0.0`      |
| `context/is_sampled/no-sdk/spec`                              | `4.7±0.05ns`         | `4.7±0.07ns`         | `0.0`      |
| `context_attach/nested_cx/empty_cx`                           | `44.6±0.31ns`        | **`19.2±0.80ns`**    | `-57`      |
| `context_attach/nested_cx/single_value_cx`                    | `44.6±0.30ns`        | **`19.6±0.42ns`**    | `-56`      |
| `context_attach/nested_cx/span_cx`                            | `44.6±0.44ns`        | **`19.5±0.20ns`**    | `-56`      |
| `context_attach/out_of_order_cx_drop/empty_cx`                | `38.5±0.40ns`        | **`20.0±0.41ns`**    | `-48`      |
| `context_attach/out_of_order_cx_drop/single_value_cx`         | `38.6±0.43ns`        | **`21.2±0.42ns`**    | `-45`      |
| `context_attach/out_of_order_cx_drop/span_cx`                 | `38.6±0.45ns`        | **`21.1±0.41ns`**    | `-45`      |
| `context_attach/single_cx/empty_cx`                           | `22.8±0.20ns`        | **`10.3±0.19ns`**    | `-55`      |
| `context_attach/single_cx/single_value_cx`                    | `22.0±0.22ns`        | **`11.0±0.20ns`**    | `-50`      |
| `context_attach/single_cx/span_cx`                            | `22.0±0.25ns`        | **`11.0±0.11ns`**    | `-50`      |
| `telemetry_suppression/enter_telemetry_suppressed_scope`      | `28.4±0.09ns`        | **`24.5±0.96ns`**    | `-14`      |
| `telemetry_suppression/is_current_telemetry_suppressed_false` | **`1.3±0.02ns`**     | `1.9±0.03ns`         | `+47`      |
| `telemetry_suppression/is_current_telemetry_suppressed_true`  | **`1.3±0.01ns`**     | `2.1±0.04ns`         | `+68`      |
| `telemetry_suppression/normal_attach`                         | `28.5±0.16ns`        | **`10.0±0.18ns`**    | `-65`      |