@bantonsson 2025-03-28
This is still a draft until #2870 has been merged and all benchmarks are run properly.